Update create account and login styling, fix cookie bug
