--- src/Components/Login-Page/Login.css
@@ -1,48 +1,0 @@
-body{
-    font-family: Georgia, 'Times New Roman', Times, serif;
-    background-color: #f4f4f4;
-    padding: 20px;
-    border-radius: 8px;
-    max-width: 400px;
-    margin: 0 auto;
-    text-align: center;
-  }
-
-  h1 {
-    color:#004990;
-  }
-  
-  h2 {
-    color: #333;
-  }
-  
-  .form-group {
-    margin-bottom: 20px;
-  }
-  
-  label {
-    margin: 5px;
-  }
-  
-  input {
-    width: 200px;
-    padding: 10px;
-    border: 1px solid #ccc;
-    border-radius: 4px;
-    box-sizing: border-box;
-    font-size: 16px;
-    align-items: center;
-    margin: 5px;
-  }
-  
-  button {
-    background-color: #004990;
-    color: white;
-    padding: 10px;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-    font-size: 16px;
-    margin: 10px;
-    font-family: Georgia, 'Times New Roman', Times, serif;
-  }
