+++ src/Components/LoginPage/Login.css
@@ -8,7 +8,7 @@ body, html {
   height: 100%;
 }
 
-.create-account-page {
+.login-page {
   text-align: center;
   display: flex;
   flex-direction: column;
@@ -39,8 +39,7 @@ label {
   color: #333;
 }
 
-input[type="text"],
-input[type="password"] {
+input {
   width: calc(100% - 20px);
   padding: 10px;
   border: 1px solid #ccc;
@@ -70,4 +69,15 @@ input[type="password"] {
 .error-message {
   color: red;
   margin: 10px 0;
+}
+
+.create-account-text {
+  color: #003366;
+  cursor: pointer;
+  margin-top: 20px;
+  font-size: 14px;
+}
+
+.create-account-text span {
+  text-decoration: underline;
 }
